+++ projects/client-api-react/src/bindings.js
@@ -42,9 +42,14 @@ const bindingsTable = {
     gravity:				[ PropTypes.oneOf(Array.from({ length: 100 }, (x, i) => i + 1)), 'defaultGravity',  undefined,  undefined],
     scalingRatio:			[ PropTypes.oneOf(Array.from({ length: 100 }, (x, i) => i + 1)), 'defaultScalingRatio', undefined,  undefined],
 
-    //togglePanel:            [ PropTypes.array,      undefined,                   undefined,      'togglePanel'],
+    setTogglePanel:            [ PropTypes.array,       undefined,                   undefined,      'togglePanel'],
     
-    ticks:                  [PropTypes.number,      undefined,                   undefined,        'tickClustering'],
+    ticks:                  [PropTypes.number,       undefined,                   undefined,      'tickClustering'],
+
+    filters:             [PropTypes.arrayOf(PropTypes.string), undefined,      undefined,      'addFilters'],
+    filter:              [PropTypes.string,       undefined,                   undefined,      'addFilter'],
+    exclusions:          [PropTypes.arrayOf(PropTypes.string), undefined,      undefined,      'addExclusions'],
+    exclusion:           [PropTypes.string,       undefined,                   undefined,      'addExclusion'],
 
     encodePointSize: [
         PropTypes.oneOfType([
@@ -118,10 +123,6 @@ const bindings =
 
 // TODO: infer this from chainList in client-api
 const calls = bindings.map(b => b.jsCommand).filter(Boolean).concat([
-    'addFilter',
-    'addFilters',
-    'addExclusion',
-    'addExclusions',
     'setSelectionExternal',
     'setHighlightExternal'
 ])