+++ CHANGELOG.md
@@ -5,6 +5,17 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## Dev
 
+## 5.0.2 - 2024-02-13
+
+### Added
+
+* **client-api**: Gracefully ignore null values passed to `togglePanel`, `addFilter`, `addExclusion`
+* **client-api**: Added `setTogglePanel`
+
+### Fix
+
+* **client-api-react**: Fix `filters`, `filter`, `exclusions`, `exclusion` to run again
+
 ## 5.0.1 - 2024-02-05
 
 ### Security